分割コンパイルをするように変更
